\name{block.obj}
\docType{data}
\alias{block.obj}
\alias{blockobj}
\title{GenABEL Object with Genotype and Phenotype Data}
\description{
  This object is of class 'gwaa.data' from the package GenABEL.  It
  contains both genotype and phenotype data in slots 'gtdata' and
  'phdata' respectively.  It is a small portion of a case-control
  genome-wide association study of Late-Onset Alzheimer's Disease
  performed at Mayo Clinic Florida.
}
\usage{block.obj}
\format{An object of class 'gwaa.data' from the package GenABEL}
\source{}
\references{Samuel Younkin, Joseph Nadeau, Robert Elston and J. Sunil
  Rao, "The Linkage Disequilibrium LASSO for SNP Selection in a Genetic
  Association Study of Late Onset Alzheimer Disease," Technical Report,
  2010\cr
Y.S. Aulchenko, S. Ripke, A. Isaacs, and C.M. van Duijn, "GenABEL: An R
Library for Genome-wide Association Analysis," Bioinformatics,
vol. 23(10), pp. 1294-1296, 2007
}
\keyword{datasets}